# openapi.yaml
openapi: "3.0.3"
info:
  title: ms-bs-practitioner-service
  description: API Contract — Dominio hce_core — Jarvis Platform
  version: "1.0.0"
paths:
...
  /TenantIdentitys:
    get:
      tags: [TenantIdentity]
      summary: Listar TenantIdentitys
      operationId: findAllTenantIdentitys
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantIdentity"
    post:
      tags: [TenantIdentity]
      summary: Crear TenantIdentity
      operationId: createTenantIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantIdentityDto"
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantIdentity"
  /TenantIdentitys/{id}:
    get:
      tags: [TenantIdentity]
      summary: Obtener TenantIdentity
      operationId: findOneTenantIdentity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantIdentity"
        "404":
          description: No encontrado
    put:
      tags: [TenantIdentity]
      summary: Actualizar TenantIdentity
      operationId: updateTenantIdentity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantIdentityDto"
      responses:
        "200":
          description: Actualizado
    delete:
      tags: [TenantIdentity]
      summary: Eliminar TenantIdentity
      operationId: deleteTenantIdentity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
...
  /Practitioners/{id}:
    get:
      tags: [Practitioner]
      summary: Obtener Practitioner
      operationId: findOnePractitioner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practitioner"
        "404":
          description: No encontrado
    put:
      tags: [Practitioner]
      summary: Actualizar Practitioner
      operationId: updatePractitioner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePractitionerDto"
      responses:
        "200":
          description: Actualizado
    delete:
      tags: [Practitioner]
      summary: Eliminar Practitioner
      operationId: deletePractitioner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
  /PractitionerRoles:
    get:
      tags: [PractitionerRole]
      summary: Listar PractitionerRoles
      operationId: findAllPractitionerRoles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PractitionerRole"
    post:
      tags: [PractitionerRole]
      summary: Crear PractitionerRole
      operationId: createPractitionerRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePractitionerRoleDto"
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PractitionerRole"
  /PractitionerRoles/{id}:
    get:
      tags: [PractitionerRole]
      summary: Obtener PractitionerRole
      operationId: findOnePractitionerRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PractitionerRole"
        "404":
          description: No encontrado
    put:
      tags: [PractitionerRole]
      summary: Actualizar PractitionerRole
      operationId: updatePractitionerRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePractitionerRoleDto"
      responses:
        "200":
          description: Actualizado
    delete:
      tags: [PractitionerRole]
      summary: Eliminar PractitionerRole
      operationId: deletePractitionerRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
...
  /PractitionerMedias:
    get:
      tags: [PractitionerMedia]
      summary: Listar PractitionerMedias
      operationId: findAllPractitionerMedias
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PractitionerMedia"
    post:
      tags: [PractitionerMedia]
      summary: Crear PractitionerMedia
      operationId: createPractitionerMedia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePractitionerMediaDto"
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PractitionerMedia"
  /PractitionerMedias/{id}:
    get:
      tags: [PractitionerMedia]
      summary: Obtener PractitionerMedia
      operationId: findOnePractitionerMedia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PractitionerMedia"
        "404":
          description: No encontrado
    put:
      tags: [PractitionerMedia]
      summary: Actualizar PractitionerMedia
      operationId: updatePractitionerMedia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePractitionerMediaDto"
      responses:
        "200":
          description: Actualizado
    delete:
      tags: [PractitionerMedia]
      summary: Eliminar PractitionerMedia
      operationId: deletePractitionerMedia
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Eliminado
...
components:
  schemas:
...
    TenantIdentity:
      type: object
      required: [tenant_identity_id, tenant_uuid, tenant_code, tenant_name, country_code, timezone, fhir_version, environment, system_version, provisioned_at, user_create, date_create, is_active]
      properties:
        tenant_identity_id:
          type: number
        tenant_uuid:
          type: string
        tenant_code:
          type: string
        tenant_name:
          type: string
        country_code:
          type: string
        timezone:
          type: string
        fhir_version:
          type: string
        environment:
          type: string
        system_version:
          type: string
        provisioned_at:
          type: string
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
    CreateTenantIdentityDto:
      type: object
      required: [tenant_identity_id, tenant_uuid, tenant_code, tenant_name, country_code, timezone, fhir_version, environment, system_version, provisioned_at, user_create, date_create, is_active]
      properties:
        tenant_identity_id:
          type: number
        tenant_uuid:
          type: string
        tenant_code:
          type: string
        tenant_name:
          type: string
        country_code:
          type: string
        timezone:
          type: string
        fhir_version:
          type: string
        environment:
          type: string
        system_version:
          type: string
        provisioned_at:
          type: string
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
...
    Practitioner:
      type: object
      required: [practitioner_id, practitioner_uuid, ad_username, name_family, name_given, active_fhir, user_create, date_create, is_active]
      properties:
        practitioner_id:
          type: number
        practitioner_uuid:
          type: string
        ad_username:
          type: string
        identifier_value:
          type: string
        identifier_system:
          type: string
        identifier_type_code:
          type: string
        name_family:
          type: string
        name_given:
          type: string
        name_prefix:
          type: string
        name_suffix:
          type: string
        gender:
          type: string
        birth_date:
          type: string
        active_fhir:
          type: boolean
        communication_language:
          type: string
        legacy_practitioner_id:
          type: string
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
    CreatePractitionerDto:
      type: object
      required: [practitioner_id, practitioner_uuid, ad_username, name_family, name_given, active_fhir, user_create, date_create, is_active]
      properties:
        practitioner_id:
          type: number
        practitioner_uuid:
          type: string
        ad_username:
          type: string
        identifier_value:
          type: string
        identifier_system:
          type: string
        identifier_type_code:
          type: string
        name_family:
          type: string
        name_given:
          type: string
        name_prefix:
          type: string
        name_suffix:
          type: string
        gender:
          type: string
        birth_date:
          type: string
        active_fhir:
          type: boolean
        communication_language:
          type: string
        legacy_practitioner_id:
          type: string
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
    PractitionerRole:
      type: object
      required: [role_id, role_uuid, practitioner_id, organisation_id, speciality_id, active_fhir, user_create, date_create, is_active]
      properties:
        role_id:
          type: number
        role_uuid:
          type: string
        practitioner_id:
          type: number
        organisation_id:
          type: number
        speciality_id:
          type: number
        location_id:
          type: number
        role_code:
          type: string
        role_display:
          type: string
        period_start:
          type: string
        period_end:
          type: string
        active_fhir:
          type: boolean
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
    CreatePractitionerRoleDto:
      type: object
      required: [role_id, role_uuid, practitioner_id, organisation_id, speciality_id, active_fhir, user_create, date_create, is_active]
      properties:
        role_id:
          type: number
        role_uuid:
          type: string
        practitioner_id:
          type: number
        organisation_id:
          type: number
        speciality_id:
          type: number
        location_id:
          type: number
        role_code:
          type: string
        role_display:
          type: string
        period_start:
          type: string
        period_end:
          type: string
        active_fhir:
          type: boolean
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
...
    PractitionerMedia:
      type: object
      required: [media_id, practitioner_id, file_server_config_id, content_type, media_file_name, relative_path, media_type, is_primary, user_create, date_create, is_active]
      properties:
        media_id:
          type: number
        practitioner_id:
          type: number
        file_server_config_id:
          type: number
        content_type:
          type: string
        media_file_name:
          type: string
        relative_path:
          type: string
        media_type:
          type: string
        media_title:
          type: string
        file_size_bytes:
          type: number
        file_hash:
          type: string
        is_primary:
          type: boolean
        attachment_date:
          type: string
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
    CreatePractitionerMediaDto:
      type: object
      required: [media_id, practitioner_id, file_server_config_id, content_type, media_file_name, relative_path, media_type, is_primary, user_create, date_create, is_active]
      properties:
        media_id:
          type: number
        practitioner_id:
          type: number
        file_server_config_id:
          type: number
        content_type:
          type: string
        media_file_name:
          type: string
        relative_path:
          type: string
        media_type:
          type: string
        media_title:
          type: string
        file_size_bytes:
          type: number
        file_hash:
          type: string
        is_primary:
          type: boolean
        attachment_date:
          type: string
        user_create:
          type: string
        user_modify:
          type: string
        date_create:
          type: string
        date_modify:
          type: string
        is_active:
          type: boolean
...